+++ test/components/naturalness/test_naturalness_analysis.py
@@ -17,6 +17,7 @@ def test_get_naturalness(operator, naturalness_utility_mock):
     polygon_geom = shapely.Polygon(((12.3, 48.2), (12.3, 48.25), (12.35, 48.25), (12.3, 48.25)))
     paths = gpd.GeoDataFrame(
         index=[1, 2],
+        data={'@other_tags': [{'tunnel': 'yes'}, {}]},
         geometry=[
             LineString([[12.4, 48.25], [12.4, 48.30]]),
             LineString([[12.41, 48.25], [12.41, 48.30]]),
@@ -25,6 +26,7 @@ def test_get_naturalness(operator, naturalness_utility_mock):
     )
     polygons = gpd.GeoDataFrame(
         index=[1, 2],
+        data={'@other_tags': [{}]},
         geometry=[polygon_geom, polygon_geom],
         crs='EPSG:4326',
     )
@@ -41,7 +43,7 @@ def test_get_naturalness(operator, naturalness_utility_mock):
             LineString([[12.4, 48.25], [12.4, 48.30]]),
             LineString([[12.41, 48.25], [12.41, 48.30]]),
         ],
-        data={'naturalness': [0.5, 0.6]},
+        data={'naturalness': [0.0, 0.6]},
         crs=CRS.from_epsg(4326),
     )